requirements.txt
# Web3 and blockchain libraries
web3>=6.11.0
solana>=0.30.2
solders>=0.18.0

# HTTP and API clients
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# CLI and formatting
rich>=13.7.0
click>=8.1.7
typer>=0.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data handling
pandas>=2.1.0

# Async support
asyncio-throttle>=1.0.2

# Logging
loguru>=0.7.2
